perf/HLIR: Move generation of the LabelMap out of the loop
Instead of generating labels on the fly, which occupied most of the time
of the loop that converts each block, I've moved that code to execute
before the loop, making the perfect case for preallocation.

I also moved from a `HashSet<usize>` for export labels to producing the
label indices in a way that export labels get the first indices so that
I just have to store an export count.

These two combined have shaved quite a path of execution, lowering the HLIR
converter to ~950ns time in the sample benchmark.